+++ recipes/api.rb
@@ -167,7 +167,7 @@ template '/etc/glance/glance-api.conf' do
   source 'glance-api.conf.erb'
   owner node['openstack']['image']['user']
   group node['openstack']['image']['group']
-  mode   00644
+  mode   00640
   variables(
     api_bind_address: api_bind.host,
     api_bind_port: api_bind.port,
@@ -205,7 +205,7 @@ template '/etc/glance/glance-cache.conf' do
   source 'glance-cache.conf.erb'
   owner node['openstack']['image']['user']
   group node['openstack']['image']['group']
-  mode   00644
+  mode   00640
   variables(
     registry_ip_address: registry_endpoint.host,
     registry_port: registry_endpoint.port,
@@ -230,7 +230,7 @@ template '/etc/glance/glance-scrubber.conf' do
   source 'glance-scrubber.conf.erb'
   owner node['openstack']['image']['user']
   group node['openstack']['image']['group']
-  mode   00644
+  mode   00640
   variables(
     registry_ip_address: registry_endpoint.host,
     registry_port: registry_endpoint.port