+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of cookbook-openstack-ima
 ## 10.0.0
 * Upgrading to Juno
 * filesystem_store_metadata_file in glance-api.conf configurable with node attribute
+* Enable rabbit_use_ssl be configurable
 
 ## 9.2.0
 * python_packages database client attributes have been migrated to the -common cookbook