+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is used to list changes made in each version of cookbook-openstack-ima
 * filesystem_store_metadata_file in glance-api.conf configurable with node attribute
 * Enable rabbit_use_ssl be configurable
 * Upgrading berkshelf from 2.0.18 to 3.1.5
+* Sync conf files with Juno
 
 ## 9.2.0
 * python_packages database client attributes have been migrated to the -common cookbook